put schengenCountries in constants.ts and use it
